+++ plugins/ftp/lib/DirList/pr_vms.cpp
@@ -131,7 +131,7 @@ BOOL net_parse_vms_dir_entry(char *line, PNET_FileEntryInfo entry_info)
      if ( *line == '[' ) {
        line++;
        e = SkipNX(line,']');
-       StrCpy( entry_info->FTPOwner, line, Min( (int)sizeof(entry_info->FTPOwner), e-line+1 ) );
+       StrCpy( entry_info->FTPOwner, line, Min( (int)sizeof(entry_info->FTPOwner), (int)(e-line+1) ) );
        line = SkipSpace( e+1 );
      }
